+++ CHANGELOG.md
@@ -4,9 +4,16 @@
 
 ### Enhancements
   - Added support for user-defined JS Hooks
+  - Added `--host` option for `mix beacon.gen.site` to serve your site at an alternative domain
 
 ### Fixes
   - Fixed a bug where MediaLibrary could check for file contents on the wrong node in multi-node deployments
+
+### Breaking Changes
+
+  - `beacon.install` - removed command aliases `-s` and `-p`
+  - `beacon.gen.site` - removed command aliases `-s` and `-p`
+  - `beacon.gen.tailwind_config` - removed command alias `-s`
 
 ## 0.3.3 (2024-12-13)